requirements.txt
discord.py
tortoise-orm
aiosqlite
asyncpg
bitstring